requirements.txt
google-adk>=1.14.0
google-cloud-aiplatform[adk,agent_engines]>=1.71.0
google-cloud-bigquery>=3.25.0
google-cloud-storage>=2.18.0
google-cloud-logging>=3.11.0
google-cloud-dlp>=3.18.0
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
opentelemetry-exporter-gcp-trace>=1.9.0
opentelemetry-exporter-gcp-monitoring>=1.9.0
pydantic>=2.8.0
pandas>=2.2.2
matplotlib>=3.9.0
python-dotenv>=1.0.1
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
httpx>=0.27.2
